feat(fleet): add traefik-prep namespace and update traefik dependencies
This commit introduces a new `traefik-prep` namespace managed by Fleet.
This namespace is intended to prepare the environment for Traefik by ensuring
the necessary namespace exists before Traefik itself is deployed.

Additionally, the `traefik` Fleet configuration has been updated to
depend on `traefik-prep` and `secrets`. This ensures that the `traefik-prep`
namespace is created and available, and that any necessary secrets are
also in place before Traefik is deployed.

The `namespace.yaml` resource has been removed from the `traefik`
kustomization as the namespace is now managed by the `traefik-prep`
Fleet bundle.