feat(cert-manager): add secretTemplate annotations to certificates for reflection
This change adds annotations to the `secretTemplate` section of the `certificate.yaml` file. These annotations enable reflection for secrets generated by cert-manager, allowing them to be mirrored into specified namespaces. This is crucial for ensuring that secrets are accessible by other services that rely on them, such as those in `cattle-system`, `media-server`, `kubernetes-dashboard`, and `traefik` namespaces.